Replace msw with miragejs
