+++ README.md
@@ -1,18 +1,18 @@
-# Vite/React/Typescript Starter
+# Vite/React/Typescript Starter (React 19 / Vite 6)
 
 A starter repository with the following libraries set up and included:
 
-- Vite
-- React
+- Vite 6
+- React 19
 - Typescript
 - Eslint
 - Prettier
 - @emotion/react
 - @emotion/styled
-- msw
+- miragejs
 - @reduxjs/toolkit
 - react-redux
-- @mui/material
+- @mui/material 7.1.0
 - notistack
 - vitest
 - @testing-library/react
