+++ src/_tests/App.spec.tsx
@@ -8,14 +8,9 @@ import { Provider } from 'react-redux';
 import { StrictMode } from 'react';
 
 import { setUpStore } from '@/store';
-import { createHandlers } from '@/_mocks/handlers';
-
 import App from '../App';
 
-import { mswTestServer } from './__setup__';
-
 function setUpApp() {
-  mswTestServer.use(...createHandlers());
   const store = setUpStore();
 
   render(
